% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directPlot.R
\name{directPlot}
\alias{directPlot}
\title{Visualize the mapping results of genomic features directly to an mRNA model.}
\usage{
directPlot(remap_results, includeNeighborDNA = TRUE, relativeProportion = c(1, 1, 1, 1),
adjust = 0.15, title  = '', legend = '', returnPlot = TRUE, type = 'relative',
trueRelativeLength = FALSE)
}
\arguments{
\item{remap_results}{A remap.results (list) object, which should be generated by \code{remapCoord} function.}

\item{includeNeighborDNA}{Whether include neighborhood regions (promoter and tail).}

\item{relativeProportion}{Ratio of the lengths of 5'UTR/CDS/3'UTR/NeighborDNA(promoter and tail) in the returned figure.}

\item{adjust}{A multiplicate bandwidth adjustment. This makes it possible to adjust the bandwidth while still using the a bandwidth estimator. For example, adjust = 1/2 means use half of the default bandwidth.}

\item{title}{Title of the returned plot.}

\item{legend}{Legend of the returned plot.}

\item{returnPlot}{If TRUE, the function returns a ggplot of genomic feature plots distribution on mRNA. If FALSE, the function returns feature density distribution values on each bin.}

\item{type}{If 'absolute', the function calculates absolute density (with the unit: number of features per bp) distribution. If 'relative', the function calculates probability density distribution.}

\item{trueRelativeLength}{If TRUE, the ratio of the lengths of each mRNA component 5'UTR/CDS/3'UTR/NeighborDNA(promoter and tail) in the returned figure follows the true length ratio among them.}
}
\value{
A figure or a data.frame object.
}
\description{
Visualize the mapping results of genomic features directly to an mRNA model. It returns the counts/width of each bin without correcting isoform ambiguity.
}
