% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoformProb.R
\name{isoformProb}
\alias{isoformProb}
\title{Calculte the probability of each genomic feature being located on different isoform transcripts.}
\usage{
isoformProb(remap_results, includeNeighborDNA = TRUE, lambda = 0)
}
\arguments{
\item{remap_results}{A remap.results (list) object, which should be generated by \code{remapCoord} function.}

\item{includeNeighborDNA}{Whether include neighborhood regions (promoter and tail).}

\item{lambda}{Penalty level for shorter transcript information. 0 means shorter transcripts are not penalized.}
}
\value{
A data frame. The columns are transcript(isoform) index, feature index, seqnames, genomic position, strand, transcript(isoform) ID, and the probability of each genomic feature being located on the corresponding isoform transcript.
}
\description{
Calculte the probability of each genomic feature being located on different isoform transcripts.
}
\examples{
file <- system.file(package="MetaTX", "extdata/remap_results_m6A.rds")
remap_results_m6A <- readRDS(file)
prob_isoforms <- isoformProb(remap_results_m6A)
}
