% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaTXplot.R
\name{metaTXplot}
\alias{metaTXplot}
\title{Visualize the density distribution of genomic features on an mRNA model.}
\usage{
metaTXplot(remap_results, includeNeighborDNA = TRUE, relativeProportion = c(1, 1, 1, 1),
lambda = 0, adjust = 0.15, title  = '', legend = '', returnPlot = TRUE, type = 'relative',
trueRelativeLength = FALSE)
}
\arguments{
\item{remap_results}{A remap.results (list) object, which should be generated by \code{remapCoord} function.}

\item{includeNeighborDNA}{Whether include neighborhood regions (promoter and tail).}

\item{relativeProportion}{Ratio of the lengths of 5'UTR/CDS/3'UTR/NeighborDNA(promoter and tail) in the returned figure.}

\item{lambda}{Penalty level for shorter transcript information. 0 means shorter transcripts are not penalized.}

\item{adjust}{A multiplicate bandwidth adjustment. This makes it possible to adjust the bandwidth while still using the a bandwidth estimator. For example, adjust = 1/2 means use half of the default bandwidth.}

\item{title}{Title of the returned plot.}

\item{legend}{Legend of the returned plot.}

\item{returnPlot}{If TRUE, the function returns a ggplot of genomic feature plots distribution on mRNA. If FALSE, the function returns feature density distribution values on each bin.}

\item{type}{If 'absolute', the function calculates absolute density (with the unit: number of features per bp) distribution. If 'relative', the function calculates probability density distribution.}

\item{trueRelativeLength}{If TRUE, the ratio of the lengths of each mRNA component 5'UTR/CDS/3'UTR/NeighborDNA(promoter and tail) in the returned figure follows the true length ratio among them.}
}
\value{
A figure or a data.frame object.
}
\description{
Visualize the density distribution of genomic features on an mRNA model. The density distribution is an EM solution that corrects isoform ambiguity among mRNAs.
}
\examples{
file <- system.file(package="MetaTX", "extdata/remap_results_m6A.rds")
remap_results_m6A <- readRDS(file)
p1 <- metaTXplot(remap_results_m6A)
}
