% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remapCoord.R
\name{remapCoord}
\alias{remapCoord}
\title{Map genomic features/regions to an mRNA model.}
\usage{
remapCoord(features,txdb = NA, num_bin = 10, includeNeighborDNA = TRUE)
}
\arguments{
\item{features}{Genomic features of interest, which should be a \code{GRanges} object. Each range should be single-based.}

\item{txdb}{A TxDb object.}

\item{num_bin}{The number of bins each mRNA component (5UTR, CDS or 3UTR) is divided into.}

\item{includeNeighborDNA}{Whether include neighborhood regions (promoter and tail).}
}
\value{
A remap.results (list) object.
\itemize{
\item \bold{\code{align_mtr}} Alignment matrix. Each row represents the mapping result of a particular feature on a particular transcript. The information of this feature and this transcript can be found on the same row in the \code{trans_info}.
\item \bold{\code{width_mtr}} Width matrix. Each row represents the width of different bins on a particular transcript.
\item \bold{\code{trans_info}} Annotation data.frame.
}
}
\description{
Map genomic features/regions to an mRNA model. An mRNA model contains three main components: 5UTR, CDS, 3UTR (probably promoter and tail). It returns a remap.results (list) object contains an alignment matrix, a width matrix and an annotation matrix.
}
\examples{
library(TxDb.Hsapiens.UCSC.hg19.knownGene)
txdb <- TxDb.Hsapiens.UCSC.hg19.knownGene
file <- system.file(package="MetaTX", "extdata/m6A_methyl.rds")
m6A_methyl<- readRDS(file)
remap_results_m6A_1 <- remapCoord(features = m6A_methyl[1:100], txdb = txdb, num_bin = 10)
}
